Refactor GPIO into according modules
